# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for product catalog management.",
        "title": "Mytheresa Catalog API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8484",
    "basePath": "/",
    "paths": {
        "/catalog": {
            "get": {
                "description": "Returns a paginated list of products, optionally filtered by category and max price.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "List products",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Pagination limit (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category code",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Filter by maximum price",
                        "name": "price_less_than",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app_catalog.catalogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mytheresa_go-hiring-challenge_app_api.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mytheresa_go-hiring-challenge_app_api.errorBody"
                        }
                    }
                }
            }
        },
        "/catalog/{code}": {
            "get": {
                "description": "Returns a single product with its variants. Variants with zero price inherit the product price.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Get product by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app_catalog.productDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_mytheresa_go-hiring-challenge_app_api.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mytheresa_go-hiring-challenge_app_api.errorBody"
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "description": "Returns all available product categories.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app_categories.categoriesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mytheresa_go-hiring-challenge_app_api.errorBody"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new product category. Both code and name are required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create category",
                "parameters": [
                    {
                        "description": "Category to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_categories.createCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app_categories.categoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mytheresa_go-hiring-challenge_app_api.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mytheresa_go-hiring-challenge_app_api.errorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app_catalog.catalogResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_catalog.productResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "app_catalog.productDetailResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_catalog.variantResponse"
                    }
                }
            }
        },
        "app_catalog.productResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "app_catalog.variantResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                }
            }
        },
        "app_categories.categoriesResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app_categories.categoryResponse"
                    }
                }
            }
        },
        "app_categories.categoryResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "app_categories.createCategoryRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_mytheresa_go-hiring-challenge_app_api.errorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        }
    }
}